+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "free-coding-models",
-  "version": "0.3.13",
+  "version": "0.3.14",
   "description": "Find the fastest coding LLM models in seconds — ping free models from multiple providers, pick the best one for OpenCode, Cursor, or any AI coding assistant.",
   "keywords": [
     "nvidia",
